Move interface to model folder.
